+++ src/components/Menu/MenuIcon.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { Icon } from 'antd'
+
+export const MenuIcon = ({ icon, className }) => {
+  if (typeof icon === 'string') {
+    return <Icon type={icon} className={className} />
+  }
+  return <div className={className}>{icon}</div>
+}