+++ frontend-core.lua
@@ -1,6 +1,7 @@
 #!/usr/bin/env tarantool
 
 local core_bundle = require('frontend-core.bundle')
+local json = require('json')
 
 local index_body = nil
 local modules = {
@@ -8,46 +9,51 @@ local modules = {
     -- [namespace] = filemap
 }
 
-local function index_handler(req)
-    if index_body == nil then
-        local entries = {}
-        for _, namespace in ipairs(modules) do
-            local mod = modules[namespace]
+local function get_index_handler(options)
+    return function(req)
+        if index_body == nil then
+            local entries = {}
+            for _, namespace in ipairs(modules) do
+                local mod = modules[namespace]
 
-            local data
-            if type(mod.__data) == 'function' then
-                data = mod.__data()
-            else
-                data = mod
-            end
+                local data
+                if type(mod.__data) == 'function' then
+                    data = mod.__data()
+                else
+                    data = mod
+                end
 
-            for filename, file in pairs(data) do
-                if file.is_entry then
-                    table.insert(entries,
-                        string.format('<script src="/static/%s/%s"></script>', namespace, filename)
-                    )
+                for filename, file in pairs(data) do
+                    if file.is_entry then
+                        table.insert(entries,
+                            string.format('<script src="'..options.prefix..'/static/%s/%s"></script>', namespace, filename)
+                        )
+                    end
                 end
             end
+
+            index_body =
+                '<!doctype html>' ..
+                '<html>' ..
+                    '<head>' ..
+                        '<title>Tarantool Cartridge</title>'..
+                        '<script>'..
+                        'window.__tarantool_admin_prefix = '..json.encode(options.prefix)..
+                        '</script>'..
+                    '</head>' ..
+                    '<body>' ..
+                        '<div id="root"></div>' ..
+                        table.concat(entries) ..
+                    '</body>' ..
+                '</html>'
         end
 
-        index_body =
-            '<!doctype html>' ..
-            '<html>' ..
-                '<head>' ..
-                    '<title>Tarantool Cartridge</title>'..
-                '</head>' ..
-                '<body>' ..
-                    '<div id="root"></div>' ..
-                    table.concat(entries) ..
-                '</body>' ..
-            '</html>'
+        return {
+            status = 200,
+            headers = {['content-type'] = 'text/html; charset=utf8'},
+            body = index_body
+        }
     end
-
-    return {
-        status = 200,
-        headers = {['content-type'] = 'text/html; charset=utf8'},
-        body = index_body
-    }
 end
 
 local function static_handler(req)
@@ -98,28 +104,42 @@ local function add(namespace, filemap)
     return true
 end
 
-local function init(httpd)
+local function init(httpd, options)
+    local options_ = {}
+
+    if options ~= nil then
+        options_ = options
+    end
+
+    local real_options = { prefix = '', enforce_root_redirect = true }
+
+    for k, v in pairs(options_) do real_options[k] = v end
+
+    local index_handler = get_index_handler(real_options)
+
     httpd:route({
-        path = '/static/:namespace/*filename',
+        path = real_options.prefix .. '/static/:namespace/*filename',
         method = 'GET',
     }, static_handler)
 
     httpd:route({
-        path = '/admin',
+        path = real_options.prefix .. '/admin',
         method = 'GET',
     }, index_handler)
 
     httpd:route({
-        path = '/admin/*any',
+        path = real_options.prefix .. '/admin/*any',
         method = 'GET',
     }, index_handler)
 
-    httpd:route({
-        path = '/',
-        method = 'GET',
-    }, function (cx)
-        return cx:redirect_to('/admin')
-    end)
+    if real_options.enforce_root_redirect then -- default true
+        httpd:route({
+            path = '/',
+            method = 'GET',
+        }, function (cx)
+            return cx:redirect_to(real_options.prefix .. '/admin')
+        end)
+    end
 
     add('core', core_bundle)
     return true