+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Customizable route prefix
+
 ## [6.3.0] - 2020-01-24
 
 - Fix side menu bug in Safari